fix 点击按钮ripple，看帖页面状态管理重构
